Migrate old api wrapper to new api decomposition
