+++ client/src/components/ResourceHub/EditResourceModal.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement, FC, Dispatch, useEffect, useState } from 'react';
 import { Modal, Button, Form, Input, Checkbox } from 'semantic-ui-react';
 import { IResource } from 'ssw-common';
 
-import { editResource, isError } from '../../utils/apiWrapper';
+import { editResource, isError } from '../../api';
 import { dbTeamToDisplay } from '../../utils/constants';
 import '../../css/EditResourceModal.css';
