+++ client/src/pages/profile/Profile.tsx
@@ -10,12 +10,7 @@ import {
 import { useParams } from 'react-router-dom';
 import { IUser } from 'ssw-common';
 
-import {
-  isError,
-  loadUser,
-  saveUser,
-  getCurrentUser,
-} from '../../utils/apiWrapper';
+import { isError, getUser, updateUser, getCurrentUser } from '../../api';
 import { teamEnum, interestsEnum, pages } from '../../utils/enums';
 import Sidebar from '../../components/Sidebar';
 import Mail from '../../assets/mail.svg';
@@ -190,7 +185,7 @@ function Profile(): ReactElement {
   }
 
   const getProfile = useCallback(async (): Promise<void> => {
-    const res = await loadUser(userId);
+    const res = await getUser(userId);
     if (isError(res)) {
       setError(true);
       setErrorMessage(res.type);
@@ -225,7 +220,7 @@ function Profile(): ReactElement {
   }, [userId]);
 
   async function updateProfile(): Promise<void> {
-    const res = await saveUser(profileData, userId);
+    const res = await updateUser(profileData, userId);
     if (isError(res)) {
       setError(true);
       setErrorMessage(res.type);
@@ -242,6 +237,8 @@ function Profile(): ReactElement {
       if (!isError(res)) {
         const currentUser: IUser = res.data.result;
         setIsEditable(userId === currentUser._id);
+      } else {
+        console.error(res);
       }
     };
